Finished updating RingsDbService to use weighted popularity scores. Popularity now has a range from 1 to 10. Changed the display of popularity and state efficiency to be a single number next to an icon.

+++ src/HallOfBeorn/Models/LotR/Simple/SimpleCard.cs
@@ -136,7 +136,7 @@ namespace HallOfBeorn.Models.LotR.Simple
         public string OctgnGuid { get; set; }
         public string RingsDbCardId { get; set; }
         public byte RingsDbPopularity { get; set; }
-        public ushort RingsDbVotes { get; set; }
+        public uint RingsDbVotes { get; set; }
 
         public List<string> Categories { get; private set; }
